[quant][graphmode] FoldConvBatchNorm2d support shared ClassTypes
Summary:
Folding Conv2d - BatchNorm2d modules means recalculate the weight and bias of Conv2d module by incorproating the parameters
of BatchNorm2d, and also change the method calls to calling only forward of Conv2d module, this involves change of both module
types and graph because the bias of Conv2d is a parameter when it has value and is an attribute when it is
None(since JIT code has assumption of prameter being Tensor in multiple places), therefore

We'll need to remove the bias attribute when it is None and add a bias parameter later. Since ClassType might be shared, we separate
remove and add in separate steps and also keep track of the processed graph to avoid modifying the graph and type multiple times.
However we'll have to record the slot index of bias as well so we can replay the slot removal on other instances of Conv2d module.

Test Plan:
tbd

Reviewers:
mvz, suo
Subscribers:

Tasks:

Tags:

ghstack-source-id: 4a543e3650096b2f09ef275d7b03d57b22ff0505
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32379